[Major Rework] Lofty Pillars (#231)
* added tut for fixing create, added mekanism and quartz

Signed-off-by: sa1672ndo <87911255+sa1672ndo@users.noreply.github.com>

* Update MODS-UNSUPPORTED.md

cleaned up some stuff and added couple of mods

Signed-off-by: sa1672ndo <87911255+sa1672ndo@users.noreply.github.com>

* Updated terms and stuff

iOS 18 is horrible

Signed-off-by: ItAnthon <145328264+ItAnthon@users.noreply.github.com>

* Update RPWORLDNOTSHOWINGUP.md

Signed-off-by: Dev1ss0 <94245667+Dev1ss0@users.noreply.github.com>

* renaming to INSTALLATIONOFMODSRPWORLDS.md

Signed-off-by: Dev1ss0 <94245667+Dev1ss0@users.noreply.github.com>

* Update config.js

Signed-off-by: Dev1ss0 <94245667+Dev1ss0@users.noreply.github.com>

* Update and rename DISCORD.md to SOCIALMEDIA.md

Signed-off-by: Dev1ss0 <94245667+Dev1ss0@users.noreply.github.com>

* Update config.js

Signed-off-by: Dev1ss0 <94245667+Dev1ss0@users.noreply.github.com>

* Updated order of FAQ platforms (hopefully works well)

Signed-off-by: Dev1ss0 <94245667+Dev1ss0@users.noreply.github.com>

* Rework into new layout as well as proper navbar utilization

This reworks the folder structure as well as the sidebar to work
alongside the navbar.

This makes mobile navigation easier.

The config file changes should allow everyone the old niceities of not
specifying the path for every bloody file while having an actual folder
structure that makes sense.

This is already a feature in vuepress v2 but oh well, here's my
reimplementation of that at least.

The new thing to edit for contributors will be arrays.js.

* Add some documentation in the form of a README.md

* Clean-up and documentation for .vuepress/config/*

* Changes to package.json and workflow to build and deploy properly

* Small corrections and changes

Edits variable names, fixes a 404 link, and reorders arrays.js to have
Android and iOS as their own objects rather than existing within FAQ in
a gigantic object property.

* [WIP] Move page-related comments and add more info to README and declutter comments

Incomplete because I'd be documenting stale things, which is pretty
stupid

* Fix indentation in config.js

* [WIP] Remove horrible hack that doesn't even work properly in github

* Remove .editorconfig and add to .gitignore

This file is supposed to be a config for an editor, hence it overrides
the configs of everyone else.

In theory this is good, in practice no one cares. This also looks
horribly ugly since it uses tabs rather than spaces. Removing it seems
sane.

* Finish rebasing on Dev1ss0/PojavlauncherWikiTesting:Dev1ss0-testing

* Indentation changes and fixes, dejankification of sidebar generation, added some jsdoc comments

Created a new function that handles non-grouped sidebar configuration in
arrays.js rather than crudely hardcode the variable names within the
sidebar definition + some indentation fixes + jsdoc syntax for comments
in js

---------

Signed-off-by: sa1672ndo <87911255+sa1672ndo@users.noreply.github.com>
Signed-off-by: ItAnthon <145328264+ItAnthon@users.noreply.github.com>
Signed-off-by: Dev1ss0 <94245667+Dev1ss0@users.noreply.github.com>
Co-authored-by: sa1672ndo <87911255+sa1672ndo@users.noreply.github.com>
Co-authored-by: ItAnthon <145328264+ItAnthon@users.noreply.github.com>
Co-authored-by: Dev1ss0 <94245667+Dev1ss0@users.noreply.github.com>