Updated iOS Versions + Replacing instructions for AltStore with SideStore.
SideStore is a fork of AltStore (This has been changed with the addition of SideStore version 0.6.x as we have branched out from using AltStore as our "core").

With SideStore, re-signing apps and installing them to your iDevice is done over VPN. A use-case for this would be sideloading/refreshing apps over Cellular without computer nearby. AltStore can only be sideloaded/re-signed if the iDevice and computer are on the same Local Wi-Fi.

Signed-off-by: 43225 <145529338+43225@users.noreply.github.com>